requirements.txt
click
GitPython
astor
python-dotenv
requests
torch              
transformers      
scikit-learn        
pydot              
graphviz              